--- packages/cli/src/core/StreamingState.ts
@@ -1,4 +1,0 @@
-export enum StreamingState {
-    Idle,
-    Responding,
-}
