--- packages/cli/src/core/constants.ts
@@ -1,1 +1,0 @@
-export const MEMORY_FILE_NAME = 'GEMINI.md';
