+++ packages/cli/src/tools/glob.tool.ts
@@ -2,8 +2,7 @@ import fs from 'fs';
 import path from 'path';
 import fg from 'fast-glob';
 import { SchemaValidator } from '../utils/schemaValidator.js';
-import { BaseTool } from './BaseTool.js';
-import { ToolResult } from './ToolResult.js';
+import { BaseTool, ToolResult } from './tool.js';
 import { shortenPath, makeRelative } from '../utils/paths.js';
 
 /**
