+++ packages/cli/src/ui/hooks/useGeminiStream.ts
@@ -1,10 +1,10 @@
 import { useState, useRef, useCallback, useEffect } from 'react';
 import { useInput } from 'ink';
-import { GeminiClient } from '../../core/GeminiClient.js';
+import { GeminiClient } from '../../core/gemini-client.js';
 import { type Chat, type PartListUnion } from '@google/genai';
 import { HistoryItem } from '../types.js';
-import { processGeminiStream } from '../../core/geminiStreamProcessor.js';
-import { StreamingState } from '../../core/StreamingState.js';
+import { processGeminiStream } from '../../core/gemini-stream.js';
+import { StreamingState } from '../../core/gemini-stream.js';
 
 const addHistoryItem = (
     setHistory: React.Dispatch<React.SetStateAction<HistoryItem[]>>,
